feat: crated custom hook for fetching restraunt manu data, created custom hook for showing online/offline status
