--- src/utils/apiDataPath.js
@@ -1,11 +1,0 @@
-// api
-//www.swiggy.com/dapi/restaurants/list/v5?lat=12.9351929&lng=77.62448069999999&page_type=DESKTOP_WEB_LISTING
-
-data.cards[4].card.card.gridElements.infoWithStyle.restaurants[0].info
-  .name;
-data.cards[4].card.card.gridElements.infoWithStyle.restaurants[0].info.cuisines;
-data.cards[4].card.card.gridElements.infoWithStyle.restaurants[0].info
-  .avgRatingString;
-data.cards[4].card.card.gridElements.infoWithStyle.restaurants[0].info.locality;
-data.cards[4].card.card.gridElements.infoWithStyle.restaurants[0].info
-  .costForTwo;
