+++ src/utils/useOnlineStatus.js
@@ -1,0 +1,20 @@
+import { useEffect, useState } from "react";
+
+const useOnlineStatus = ()=> {
+    const [onlineStatus, setOnlineStatus] = useState(true);
+    // check whether we are online or offline
+    useEffect(()=> {
+        window.addEventListener('offline', ()=> {
+            setOnlineStatus(false)
+        })
+
+        window.addEventListener('online', ()=> {
+            setOnlineStatus(true)
+        })
+    },[])
+
+    // boolean value
+    return onlineStatus;
+}
+
+export default useOnlineStatus;
